requirements.txt
fastapi==0.95.0
openai==0.27.2
pandas==1.5.3
pydantic==1.10.6
uvicorn==0.21.1
Jinja2==3.1.2
langchain==0.0.119
gpt-index==0.4.35.post1
fastapi_sqlalchemy==0.2.1
scipy==1.9.3
discord==2.2.2
pillow==9.4.0
google-auth==2.17.0
google-cloud-storage==2.8.0
git+https://github.com/alex200420/brainbite.git@main#egg=brainbite